+++ trin-validation/Cargo.toml
@@ -22,7 +22,9 @@ ethportal-api = { path = "../ethportal-api" }
 lazy_static = "1.4.0"
 serde = { version = "1.0.150", features = ["derive"] }
 serde_json = "1.0.89"
+surf = "2.3.2"
 tokio = { version = "1.14.0", features = ["full"] }
+tracing = "0.1.36"
 tree_hash = "0.4.0"
 tree_hash_derive = "0.4.0"
 trin-types = { path = "../trin-types" }
