+++ trin-validation/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod accumulator;
+pub mod bridge;
 pub mod constants;
 pub mod merkle;
 pub mod oracle;
